# gpu_hist performance fixes