There are a few fixes here:
 - #5534 introduced a performance regression. Caching device vectors were used as temporary vectors in the EvaluateSplit code, creation of these vectors introduces synchronisation, slowing down the algorithm. This PR introduces TemporaryArray as a simple data structure, still using the same allocator internally, but avoiding the synchronisation overhead.
- Remove calls to cudaSetDevice on a per node basis. This function is called over 100,000 times when building 500 trees, which is a little excessive. While it is relatively cheap, this does have a noticeable impact. I have removed all calls during each node construction, but left calls that occur only once per tree.
- Fixed a bug where the histogram storage would double on every call up to `size_t kStopGrowingSize = 1 << 26`, which allocates about 2GB of memory, regardless of how much memory is actually needed. This saves a huge amount of working memory and also makes the algorithm faster because we don't need to zero 2GB of memory every boosting iteration.

Running `python benchmark_tree.py` on my windows system with a 1080Ti, 500 iterations with 10M*50 training instances, training time goes from 18s to 10.5s. If we consider before the regression of #5534 it goes from 12.5s to 10.5s, an improvement of about 1.2x.